Fixed bug with chained if instructions.
When evaluating an if instruction, the program counter would be incremented to skip the next
instruction, then that instruction was checked to see if it was an if that should be skipped.
Instead, it should check the next instruction to see if it is an if, then increment to skip
the "protected" instruction. Essentially it was checking the second instruction for chaining
instead of the first.

Fixed by switching the order of evaluation so that if chains are skipped, and then the final
instruction they were "protecting" is skipped.